# docs/openapi/CampusHelpATW.openapi.yaml
openapi: 3.0.3
info:
  title: CampusHelpATW API
  version: 0.1.0
  description: |
    CampusHelpATW 微信小程序后端接口（FastAPI 实现基线）。
    实时聊天 WebSocket 端点：
    - `/api/v1/ws/tasks/{task_id}?token={jwt}`
    全局未读通知 WebSocket 端点：
    - `/api/v1/ws/me/notifications?token={jwt}`
    全局通道事件：
    - `unread_snapshot` / `chat_unread`（聊天未读）
    - `task_event`（任务关键节点）
servers:
  - url: http://127.0.0.1:3000
    description: Local
  - url: https://api.example.com
    description: Production (placeholder)
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/tasks:
    get:
      tags: [tasks]
      summary: 任务列表
      description: 返回任务列表，并附带 `publisher_user` / `acceptor_user` 简要信息。
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
        - in: query
          name: include_unread
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: 列表结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [tasks]
      summary: 创建任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '200':
          description: 创建结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/tasks/{task_id}:
    get:
      tags: [tasks]
      summary: 任务详情
      description: 返回任务信息，并附带 `publisher_user` / `acceptor_user` 简要信息。
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
        - in: query
          name: include_unread
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: 详情结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/tasks/{task_id}/status-logs:
    get:
      tags: [tasks]
      summary: 任务状态日志
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 日志列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/tasks/{task_id}/accept:
    post:
      tags: [tasks]
      summary: 接单
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskActionReasonRequest'
      responses:
        '200':
          description: 接单结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/tasks/{task_id}/start:
    post:
      tags: [tasks]
      summary: 标记进行中
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskActionReasonRequest'
      responses:
        '200':
          description: 更新结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/tasks/{task_id}/submit-completion:
    post:
      tags: [tasks]
      summary: 提交完成
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskActionReasonRequest'
      responses:
        '200':
          description: 更新结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/tasks/{task_id}/confirm-completion:
    post:
      tags: [tasks]
      summary: 确认完成
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskActionReasonRequest'
      responses:
        '200':
          description: 更新结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/tasks/{task_id}/dispute:
    post:
      tags: [tasks]
      summary: 发起申诉
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskActionReasonRequest'
      responses:
        '200':
          description: 更新结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/tasks/{task_id}/cancel:
    post:
      tags: [tasks]
      summary: 取消任务
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskActionReasonRequest'
      responses:
        '200':
          description: 更新结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/tasks/{task_id}/reviews:
    post:
      tags: [tasks]
      summary: 提交任务评价
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskReviewRequest'
      responses:
        '200':
          description: 评价结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/users/{user_id}:
    get:
      tags: [tasks]
      summary: 查询用户公开资料
      description: 返回基础资料、关系状态、评价聚合以及共同数据聚合。
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 用户公开资料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/users/{user_id}/reviews:
    get:
      tags: [tasks]
      summary: 查询用户评价列表
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
        - in: query
          name: rating
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - in: query
          name: with_content
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 评价列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/users/{user_id}/shared/tasks:
    get:
      tags: [tasks]
      summary: 查询共同任务列表
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
        - in: query
          name: status
          schema:
            type: string
            enum:
              - OPEN
              - ACCEPTED
              - IN_PROGRESS
              - PENDING_CONFIRM
              - DONE
              - CANCELED
              - DISPUTED
        - in: query
          name: sort
          schema:
            type: string
            default: latest
            enum:
              - latest
              - deadline_asc
              - reward_desc
      responses:
        '200':
          description: 共同任务列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/me/points/ledger:
    get:
      tags: [tasks]
      summary: 查询我的积分流水
      parameters:
        - in: query
          name: point_type
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 积分流水列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
      required: [code, message]
...
    TaskCreateRequest:
      type: object
      required: [title, description, category, deadline_at]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        location_text:
          type: string
        reward_amount:
          type: number
        reward_type:
          type: string
        deadline_at:
          type: string
          format: date-time
    TaskActionReasonRequest:
      type: object
      properties:
        reason:
          type: string
    TaskReviewRequest:
      type: object
      required: [reviewee_id, rating]
      properties:
        reviewee_id:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        content:
          type: string
...